# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Endpoints do backoffice RCM Tech.",
        "title": "RCM Backoffice API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/commissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Lista comissoes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/finance.Commission"
                            }
                        }
                    }
                }
            }
        },
        "/commissions/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Aprova comissao",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
        "/contracts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Lista contratos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/contract.Contract"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Cria contrato",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/contract.Contract"
                        }
                    }
                }
            }
        },
        "/contracts/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Atualiza contrato",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Remove contrato",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Lista clientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/customer.Customer"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Cria cliente",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/customer.Customer"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Atualiza cliente",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Remove cliente",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Autentica usuario",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    }
                }
            }
        },
        "/promoters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "promoters"
                ],
                "summary": "Lista promotores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/promoter.Promoter"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "promoters"
                ],
                "summary": "Cria promotor",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/promoter.Promoter"
                        }
                    }
                }
            }
        },
        "/promoters/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "promoters"
                ],
                "summary": "Atualiza promotor",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "promoters"
                ],
                "summary": "Remove promotor",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
        "/receivables": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Lista contas a receber",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/finance.AccountReceivable"
                            }
                        }
                    }
                }
            }
        },
        "/receivables/{id}/pay": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Marca receivable como pago",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        },
        "/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Lista servicos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.Service"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Cria servico",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service.Service"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Atualiza servico",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "services"
                ],
                "summary": "Remove servico",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "null"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "contract.Contract": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "promoter_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value_total": {
                    "type": "number"
                }
            }
        },
        "customer.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "documentID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "legalName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "promoterID": {
                    "type": "string"
                },
                "tradeName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "finance.AccountReceivable": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contractID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paidAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "finance.Commission": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "approved": {
                    "type": "boolean"
                },
                "approvedAt": {
                    "type": "string"
                },
                "approvedBy": {
                    "type": "string"
                },
                "contractID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "promoterID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "promoter.Promoter": {
            "type": "object",
            "properties": {
                "bankAccount": {
                    "type": "object"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "documentID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "service.Service": {
            "type": "object",
            "properties": {
                "basePrice": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}